requirements_python312.txt
# Core Framework - Python 3.12 Compatible
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Advanced ML & AI - Updated for Python 3.12
transformers==4.36.0
torch==2.2.0
accelerate==0.25.0
bitsandbytes==0.41.3
optimum==1.16.1

# Vector Database & Embeddings
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.18
qdrant-client==1.7.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
numpy>=1.26.0
polars==0.20.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
markdown==3.5.1
beautifulsoup4==4.12.2
langchain==0.1.0
langchain-community==0.0.10

# Caching & Performance
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Async & Concurrency
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Configuration
python-dotenv==1.0.0
hydra-core==1.3.2

# Additional Python 3.12 compatibility fixes
setuptools>=68.0.0
wheel>=0.40.0
